rename 'constructor' to 'construct' allow nyc to function
