+++ lib/construct/bffs.js
