+++ lib/construct/fitting.js
