+++ lib/construct/index.js
