+++ lib/construct/npm/index.js
