+++ lib/construct/progress.js
@@ -134,7 +134,7 @@ Progress.prototype.done = function done(id) {
 Progress.prototype.ignore = function ignore() {
   this.fn();
   this.cleanup();
-  this._endStatus({ message: 'ignored' });
+  this._endStatus('ignored');
   return this.write({
     event: 'task',
     message: 'ignored',
@@ -182,7 +182,7 @@ Progress.prototype._writeStatus = function writeStatus(data) {
 
   const { name, env, version, locale, type: buildType } = this.metadata || {};
   const freshSpec = {
-    eventType: 'status-event',
+    eventType: 'event',
     name,
     env,
     version,
@@ -193,14 +193,14 @@ Progress.prototype._writeStatus = function writeStatus(data) {
   nsqStream.write(freshSpec);
 };
 
-Progress.prototype._endStatus = function endStatus(err) {
+Progress.prototype._endStatus = function endStatus(type, err) {
   if (!this.nsqStream || this.nsqStream._writableState.ended) {
     return;
   }
 
   const { name, env, version, type: buildType } = this.metadata || {};
   const freshSpec = {
-    eventType: 'status',
+    eventType: type,
     name,
     env,
     version,
@@ -239,7 +239,7 @@ Progress.prototype.end = function end(err) {
   });
 
   this.stream.end();
-  this._endStatus(err);
+  this._endStatus(err ? 'error' : 'queued', err);
 
   return this;
 };
