+++ lib/preboots/index.js
@@ -27,9 +27,9 @@ module.exports = function preboot(app, options, next) {
   app.preboot(require('./cdnup'));
   app.preboot(require('./nsq.js'));
   app.preboot(require('./npm'));
-  app.preboot(require('../constructor/bffs'));
+  app.preboot(require('../construct/bffs'));
   app.preboot(require('./scheduler'));
-  app.preboot(require('../constructor'));
+  app.preboot(require('../construct'));
   app.preboot(require('./terminate'));
   app.preboot(require('./feedsme'));
