+++ package.json
@@ -85,7 +85,7 @@
     "eslint-plugin-json": "^1.2.0",
     "eslint-plugin-mocha": "^4.11.0",
     "hyperquest": "~1.2.0",
-    "mocha": "~3.1.0",
+    "mocha": "^5.2.0",
     "nock": "2.10.0",
     "nyc": "^11.8.0",
     "sinon": "^3.2.1"
